+++ lib/quiz_view.dart
@@ -1,4 +1,5 @@
 library quiz_view;
+
 import 'package:flutter/material.dart';
 import 'dart:math';
 
@@ -38,65 +39,12 @@ class QuizView extends StatefulWidget {
       @required this.onRightAnswer,
       @required this.onWrongAnswer});
 
-  _QuizViewState createState() => _QuizViewState(
-      this.showCorrect,
-      this.questionTag,
-      this.question,
-      this.questionColor,
-      this.backgroundColor,
-      this.image,
-      this.height,
-      this.width,
-      this.rightAnswer,
-      this.wrongAnswers,
-      this.tagColor,
-      this.tagBackgroundColor,
-      this.answerColor,
-      this.answerBackgroundColor,
-      this.onRightAnswer,
-      this.onWrongAnswer);
+  _QuizViewState createState() => _QuizViewState();
 }
 
-class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
-  @override
-  void iniState() {
-    super.initState();
-  }
+class _QuizViewState extends State<QuizView> {
+  _QuizViewState();
 
-  _QuizViewState(
-      this.showCorrect,
-      this.questionTag,
-      this.question,
-      this.questionColor,
-      this.backgroundColor,
-      this.image,
-      this.height,
-      this.width,
-      this.rightAnswer,
-      this.wrongAnswers,
-      this.tagColor,
-      this.tagBackgroundColor,
-      this.answerColor,
-      this.answerBackgroundColor,
-      this.onRightAnswer,
-      this.onWrongAnswer);
-
-  bool showCorrect;
-  String question;
-  String questionTag;
-  Color questionColor;
-  Color backgroundColor;
-  Widget image;
-  double width;
-  double height;
-  String rightAnswer;
-  List<String> wrongAnswers;
-  Color tagBackgroundColor;
-  Color tagColor;
-  Color answerColor;
-  Color answerBackgroundColor;
-  void Function() onRightAnswer;
-  void Function() onWrongAnswer;
   bool isTapped = false;
   static Random _random = new Random();
   int answerIndex;
@@ -104,7 +52,7 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
     if (!isTapped) {
-      answerIndex = _random.nextInt(wrongAnswers.length + 1);
+      answerIndex = _random.nextInt(widget.wrongAnswers.length + 1);
     }
 
     Column answerColumn = Column(
@@ -113,27 +61,29 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
       children: [],
     );
 
-    for (String i in wrongAnswers) {
+    for (String i in widget.wrongAnswers) {
       answerColumn.children.add(Container(
         width: MediaQuery.of(context).size.width,
         padding: EdgeInsets.all(15),
         child: RaisedButton(
           padding: EdgeInsets.fromLTRB(20, 10, 20, 10),
           shape: StadiumBorder(),
-          color: answerBackgroundColor,
+          color: widget.answerBackgroundColor,
           textColor: Colors.white,
           child: Center(
             child: Text(
               i,
               style: TextStyle(
-                  color: answerColor,
-                  fontSize: width > height ? width / 25 : height / 25),
+                  color: widget.answerColor,
+                  fontSize: widget.width > widget.height
+                      ? widget.width / 25
+                      : widget.height / 25),
             ),
           ),
           onPressed: () {
             if (!isTapped) {
-              onWrongAnswer();
-              if (showCorrect) {
+              widget.onWrongAnswer();
+              if (widget.showCorrect) {
                 setState(() {
                   isTapped = !isTapped;
                 });
@@ -151,19 +101,21 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
           child: RaisedButton(
             padding: EdgeInsets.fromLTRB(20, 10, 20, 10),
             shape: StadiumBorder(),
-            color: isTapped ? Colors.green : answerBackgroundColor,
+            color: isTapped ? Colors.green : widget.answerBackgroundColor,
             child: Center(
               child: Text(
-                rightAnswer,
+                widget.rightAnswer,
                 style: TextStyle(
-                    color: answerColor,
-                    fontSize: width > height ? width / 25 : height / 25),
+                    color: widget.answerColor,
+                    fontSize: widget.width > widget.height
+                        ? widget.width / 25
+                        : widget.height / 25),
               ),
             ),
             onPressed: () {
               if (!isTapped) {
-                onRightAnswer();
-                if (showCorrect) {
+                widget.onRightAnswer();
+                if (widget.showCorrect) {
                   setState(() {
                     isTapped = !isTapped;
                   });
@@ -175,14 +127,16 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
 
     return Container(
         decoration: BoxDecoration(
-            color: backgroundColor,
-            borderRadius: BorderRadius.all(
-                Radius.circular(height > width ? height / 20 : width / 20)),
+            color: widget.backgroundColor,
+            borderRadius: BorderRadius.all(Radius.circular(
+                widget.height > widget.width
+                    ? widget.height / 20
+                    : widget.width / 20)),
             border: Border.all(
                 color: Colors.black, width: 1.5, style: BorderStyle.solid)),
-        width: width,
-        height: height,
-        child: questionTag != null
+        width: widget.width,
+        height: widget.height,
+        child: widget.questionTag != null
             ? Stack(
                 children: [
                   Positioned(
@@ -191,46 +145,56 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
                     child: Container(
                       padding: EdgeInsets.all(10),
                       decoration: BoxDecoration(
-                          color: tagBackgroundColor,
+                          color: widget.tagBackgroundColor,
                           borderRadius: BorderRadius.only(
                             topLeft: Radius.circular(
-                                height > width ? height / 20 : width / 20),
+                                widget.height > widget.width
+                                    ? widget.height / 20
+                                    : widget.width / 20),
                             bottomRight: Radius.circular(
-                                height > width ? height / 20 : width / 20),
+                                widget.height > widget.width
+                                    ? widget.height / 20
+                                    : widget.width / 20),
                           )),
                       child: Text(
-                        questionTag,
+                        widget.questionTag,
                         style: TextStyle(
-                            color: tagColor,
+                            color: widget.tagColor,
                             fontWeight: FontWeight.bold,
-                            fontSize:
-                                width > height ? width / 20 : height / 20),
+                            fontSize: widget.width > widget.height
+                                ? widget.width / 20
+                                : widget.height / 20),
                       ),
                     ),
                   ),
                   Container(
                     margin: EdgeInsets.fromLTRB(
-                        0, width > height ? width / 10 : height / 10, 0, 0),
+                        0,
+                        widget.width > widget.height
+                            ? widget.width / 10
+                            : widget.height / 10,
+                        0,
+                        0),
                     child: SingleChildScrollView(
                       child: Column(
                         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                         crossAxisAlignment: CrossAxisAlignment.center,
-                        children: image != null
+                        children: widget.image != null
                             ? [
                                 Container(
                                   padding: EdgeInsets.all(20),
                                   child: RichText(
                                       text: TextSpan(
-                                    text: question,
+                                    text: widget.question,
                                     style: TextStyle(
-                                        color: questionColor,
+                                        color: widget.questionColor,
                                         fontWeight: FontWeight.bold,
-                                        fontSize: width > height
-                                            ? width / 20
-                                            : height / 20),
+                                        fontSize: widget.width > widget.height
+                                            ? widget.width / 20
+                                            : widget.height / 20),
                                   )),
                                 ),
-                                image,
+                                widget.image,
                                 answerColumn
                               ]
                             : [
@@ -238,13 +202,14 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
                                   padding: EdgeInsets.all(20),
                                   child: RichText(
                                     text: TextSpan(
-                                        text: question,
+                                        text: widget.question,
                                         style: TextStyle(
-                                            color: questionColor,
+                                            color: widget.questionColor,
                                             fontWeight: FontWeight.bold,
-                                            fontSize: width > height
-                                                ? width / 20
-                                                : height / 20)),
+                                            fontSize:
+                                                widget.width > widget.height
+                                                    ? widget.width / 20
+                                                    : widget.height / 20)),
                                   ),
                                 ),
                                 answerColumn
@@ -258,22 +223,22 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
                 child: Column(
                   mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                   crossAxisAlignment: CrossAxisAlignment.center,
-                  children: image != null
+                  children: widget.image != null
                       ? [
                           Container(
                             padding: EdgeInsets.all(20),
                             child: RichText(
                                 text: TextSpan(
-                              text: question,
+                              text: widget.question,
                               style: TextStyle(
-                                  color: questionColor,
+                                  color: widget.questionColor,
                                   fontWeight: FontWeight.bold,
-                                  fontSize: width > height
-                                      ? width / 20
-                                      : height / 20),
+                                  fontSize: widget.width > widget.height
+                                      ? widget.width / 20
+                                      : widget.height / 20),
                             )),
                           ),
-                          image,
+                          widget.image,
                           answerColumn
                         ]
                       : [
@@ -281,13 +246,13 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
                             padding: EdgeInsets.all(20),
                             child: RichText(
                               text: TextSpan(
-                                  text: question,
+                                  text: widget.question,
                                   style: TextStyle(
-                                      color: questionColor,
+                                      color: widget.questionColor,
                                       fontWeight: FontWeight.bold,
-                                      fontSize: width > height
-                                          ? width / 20
-                                          : height / 20)),
+                                      fontSize: widget.width > widget.height
+                                          ? widget.width / 20
+                                          : widget.height / 20)),
                             ),
                           ),
                           answerColumn
@@ -296,4 +261,3 @@ class _QuizViewState extends State<QuizView> with TickerProviderStateMixin {
               ));
   }
 }
-