+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [0.0.1] - Release
+
+* A customizable quiz view widget.
+
 ## [0.0.2] - Release
+
+* A customizable quiz view widget.
+
+## [0.0.3] - Release
+
+* A customizable quiz view widget.
+
+## [0.0.4] - Release
+
+* A customizable quiz view widget.
+
+## [0.0.5] - Release
+
+* A customizable quiz view widget.
+
+## [1.0.0] - Release
 
 * A customizable quiz view widget.